// backend/src/co-tuc/dto/request/create-chi-tiet-co-tuc.request.dto.ts
export class CreateChiTietCoTucRequestDto {
    soDKSH: string;

    soDkshNamChot: string;

    slckngChuaLk: number;

    slckngDaLk: number;

    slckngChuaLkTruocTax: number;

    slckngDaLkTruocTax: number;

    slckngChuaLkSauTax: number;

    slckngDaLkSauTax: number;

    thanhToanChuaLk: boolean;

    thanhToanDaLk: boolean;

    namChot: number;
}